% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Creates a leaflet map of earthquake locations}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{NOAA earthquake data to be displayed}

\item{annot_col}{Name of the column to be displayed in the popups. Defaults to `"DATE"`.}
}
\value{
Creates a leaflet map of earthquake epicenters with annotation popups.
}
\description{
Function that creates a leaflet map of earthquake epicenters with popups containing information on a selected
column of the utilised earthquake data.
Function to be used with NOAA earthquake data downloaded and cleaned with the \code{\link{eq_read_data}} and
\code{\link{eq_clean_data}} functions of the `earthquakesAJS`` package.
}
