% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Draw a timeline of NOAA earthquake events}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created}

\item{data}{Input data for timeline}

\item{stat}{The statistical transformation to use on the data}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment
function}

\item{na.rm}{If `FALSE`, the default, missing values are removed with a warning. If `TRUE`,
missing values are silently removed}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{If `FALSE``, overrides the default aesthetics, rather than
combining with them}

\item{...}{Other arguments passed on to `[layer()]`. These are often aesthetics, used to set an aesthetic
to a fixed value, like `colour ="blue"`}
}
\value{
Draws a timeline of earthquake events for one or more selected
 countries
}
\description{
Function that draws a timeline of earthquake events for one or more selected
countries. The timeline can optionally feature the magnitude and number of
victimes of earthquakes represented by points of different size and/or colour.
Function to be used in a \code{\link{ggplot}} call to NOAA earthquake data downloaded and cleaned with the
\code{\link{eq_read_data}} and \code{\link{eq_clean_data}} functions of the `earthquakesAJS`` package.
}
\examples{
\dontrun{
url<-"https://www.example-url.org"
ds <- eq_location_clean(eq_clean_data(ds))
aux<-dplyr::filter(ds, DATE >= "1980-01-01" & DATE <="2014-01-01" & COUNTRY == c("JAPAN","CHILE"))
ggplot2::ggplot(aux) +
geom_timeline(ggplot2::aes(x=DATE, y=COUNTRY, size=EQ_MAG_MB, colour=DEATHS)) +
viridis::scale_colour_viridis(option="D") +
ggplot2::theme_minimal()
}

}
