% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Timeline object for the geom_timeline function.}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\description{
\code{ggproto} object underlying the \code{\link{geom_timeline}} function of the `earthquakesAJS`` package.
}
\keyword{datasets}
