% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Timeline label object for the geom_timeline_label function.}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
GeomTimelineLabel
}
\description{
\code{ggproto} object underlying the \code{\link{geom_timeline_label}} function of the `earthquakesAJS` package.
}
\keyword{datasets}
