% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Add more information to map popups}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{NOAA earthquake data from which to exract information}
}
\value{
A HTML label containing information of the location, magnitude and severity (total deaths) of
earthquakes, if corresponding information is present in the input data
}
\description{
Function that adds additional information to popups creates with the \code{\link{eq_map}} function of the
`earthquakesAJS`` package.
Function to be used with NOAA earthquake data downloaded and cleaned with the \code{\link{eq_read_data}} and
\code{\link{eq_clean_data}} functions of the `earthquakesAJS`` package.
}
