% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_tidy.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Reformat location names in NOAA earthquake data}
\usage{
eq_location_clean(data)
}
\value{
Returns a table with a cleaned "LOCATION_NAME" column where the country name has been stripped
and location names have been converted to title case
}
\description{
Function that reformats location names in NOAA earthquake data.
}
